fix&feat: rewrite spawn and improve imports
