+++ src/commands/arch.js
@@ -2,11 +2,11 @@
  * Discord Bot Arch Command
  * @module arch
  */
-const PATH = require('path');
+const PATH = require('node:path');
 const { SlashCommandBuilder, EmbedBuilder, AttachmentBuilder, ButtonBuilder, ActionRowBuilder, ButtonStyle } = require('discord.js');
-const AUR = require(PATH.join(__dirname, '..', 'modules', 'callAUR.js'));
-const SPAWN = require(PATH.join(__dirname, '..', 'modules', 'spawn.js'));
-const LOGO = new AttachmentBuilder(PATH.join(__dirname, '..', 'assets', 'archlinux.png'));
+const AUR = require('#modules/callAUR.js');
+const SPAWN = require('#modules/fetch-pkgs/index.js');
+const LOGO = new AttachmentBuilder(PATH.join(__dirname, '..', '..', 'assets', 'archlinux.png'));
 const ROW = new ActionRowBuilder();
 let embed = new EmbedBuilder();
 const STORED = 'stored';
@@ -179,26 +179,31 @@ module.exports = {
 		.setDescription('Performs a search in the Arch package database for packages')
 		.setDescriptionLocalizations({
 			'sv-SE': 'Utför en sökning i Archs paketförråd efter paket',
+			'es-ES': 'Hace una busqueda de paquetes en la base de datos de Arch'
 		})
 		.addStringOption(option => option
 			.setName('package')
-			.setDescription('The package you\'re search for')
+			.setDescription('The package you\'re searching for')
 			.setRequired(true)
 			.setNameLocalizations({
 				'sv-SE': 'paket',
+				'es-ES': 'paquete'
 			})
 			.setDescriptionLocalizations({
 				'sv-SE': 'Paketet du vill söka efter',
+				'es-ES': 'El paquete que buscas'
 			}),
 		)
 		.addStringOption(option => option
 			.setName('repository')
 			.setNameLocalizations({
 				'sv-SE': 'paketförråd',
+				'es-ES': 'repositorio'
 			})
 			.setDescription('Which repository to fetch from')
 			.setDescriptionLocalizations({
 				'sv-SE': 'Vilket paketförråd att söka igenom',
+				'es-ES': 'De que repositorio buscar'
 			})
 			.addChoices(
 				{ name: 'Core', value: 'core' },
@@ -215,20 +220,24 @@ module.exports = {
 			.setName('dependencies')
 			.setNameLocalizations({
 				'sv-SE': 'beroenden',
+				'es-ES': 'dependencias'
 			})
 			.setDescription('Whether or not to show dependencies for the given package')
 			.setDescriptionLocalizations({
 				'sv-SE': 'Huruvida beroenden för det givna paketet ska visas eller inte',
+				'es-ES': 'Si mostrar o no las dependencias del paquete'
 			}),
 		)
 		.addBooleanOption(option => option
 			.setName('ephemeral')
 			.setNameLocalizations({
 				'sv-SE': 'efemär',
+				'es-ES': 'efímero'
 			})
 			.setDescription('Toggles whether or not this message should be ephemeral')
 			.setDescriptionLocalizations({
 				'sv-SE': 'Växlar om detta meddelande ska vara kortlivat eller inte',
+				'es-ES': 'Establece si este mensaje debería ser efímero'
 			}),
 		),
 
@@ -267,7 +276,7 @@ module.exports = {
 
 		}
 
-		const DIRECTORY = await readdir(STORED);
+		const DIRECTORY = await readdir(PATH.join(__dirname, '..', STORED));
 		const REPOSITORIES = DIRECTORY.filter(file => file.endsWith('.json'));
 
 		const PACKAGES = [];
@@ -345,6 +354,7 @@ module.exports = {
 				if (DEPENDENCIES) embed.setDescription(`Please select which package you want to show.\u2007\n${description}`);
 				else embed.setDescription(`Please select which package you want to show.\n${description}`);
 
+				// ! This line causes discord.js to use `buffer.Blob`, which is considered a experimental feature in this node version
 				await Interaction.editReply({ content: '', embeds: [embed], files: [LOGO], components: [ROW] });
 
 			}
